Mine will produce consumables that are reflected int he HUD

+++ src/org/rubenrr/walkeitor/config/ElementConfig.java
@@ -157,7 +157,7 @@ public enum ElementConfig {
         return production.getConsumableProduce();
     }
 
-    public int getTimeElapsed() {
+    public float getTimeElapsed() {
         return production.getTimeElapsed();
     }
